PYTHON/23-24_BD2_TP2_GuilhermeGonçalves_Hugo-Pais/requirements.txt
Flask
PyJWT
psycopg2-binary
python-dotenv
